Default to 'info' command if subcommand not found.
A lot of times I just want information about a package such as who
owns it and I try to do `pkgwat <the package>` forgetting that I need
`info` before the name. I might be the only one who forgets to do this,
but the patch seemed noninvasive so I decided to make it default to
info.

So now you can do `pkgwat firefox` or just type `firefox` in the pkgwat
REPL. Other commands still work fine, and `info` still works if
specified.