remove custom override of stdout/stderr encoding
The override is needed when piping stream output instead of printing to
terminal, and just with Python 2. With Python 3, this causes an error:
```
TypeError: write() argument must be str, not bytes
```

Fortunately the cliff library now handles this internally
out-of-the-box (just for Python 2), so we don't need to do this at all.

The correct functionality can be tested for example with:
```
$ pkgwat info aesh | grep description
```
(notice `Æsh` in description)